+++ tests/supernotify/test_envelope.py
@@ -1,11 +1,6 @@
 import time
-from unittest.mock import Mock
 
-from custom_components.supernotify import (
-    DELIVERY_SELECTION_IMPLICIT,
-)
-from custom_components.supernotify.context import Context
-from custom_components.supernotify.delivery import Delivery
+from custom_components.supernotify import CONF_TRANSPORT
 from custom_components.supernotify.envelope import Envelope
 from custom_components.supernotify.model import MessageOnlyPolicy
 from custom_components.supernotify.notification import Notification
@@ -63,31 +58,31 @@ async def test_repr() -> None:
     assert repr(envelope) == "Envelope(message=Hello Test,title=None,delivery=DEFAULT_notify_entity)"
 
 
-def test_message_usage(mock_context: Context) -> None:
-    delivery = Mock(spec=Delivery, title=None, message=None, selection=DELIVERY_SELECTION_IMPLICIT)
-    delivery.name = "push"
-    mock_context.delivery_registry.deliveries = {"push": delivery}
+async def test_message_usage() -> None:
+    ctx = TestingContext(deliveries={"push": {CONF_TRANSPORT: "notify_entity"}})
+    await ctx.test_initialize()
+    delivery = ctx.delivery("push")
 
-    uut = Envelope(delivery, Notification(mock_context, "testing 123", title="the big title"))
+    uut = Envelope(delivery, Notification(ctx, "testing 123", title="the big title"))
     assert uut._compute_message() == "testing 123"
     assert uut._compute_title() == "the big title"
 
-    delivery.option_str.return_value = MessageOnlyPolicy.USE_TITLE
-    uut = Envelope(delivery, Notification(mock_context, "testing 123", title="the big title"))
+    delivery.options["message_usage"] = MessageOnlyPolicy.USE_TITLE
+    uut = Envelope(delivery, Notification(ctx, "testing 123", title="the big title"))
     assert uut._compute_message() == "the big title"
     assert uut._compute_title() is None
 
-    delivery.option_str.return_value = MessageOnlyPolicy.USE_TITLE
-    uut = Envelope(delivery, Notification(mock_context, "testing 123"))
+    delivery.options["message_usage"] = MessageOnlyPolicy.USE_TITLE
+    uut = Envelope(delivery, Notification(ctx, "testing 123"))
     assert uut._compute_message() == "testing 123"
     assert uut._compute_title() is None
 
-    delivery.option_str.return_value = MessageOnlyPolicy.COMBINE_TITLE
-    uut = Envelope(delivery, Notification(mock_context, "testing 123", title="the big title"))
+    delivery.options["message_usage"] = MessageOnlyPolicy.COMBINE_TITLE
+    uut = Envelope(delivery, Notification(ctx, "testing 123", title="the big title"))
     assert uut._compute_message() == "the big title testing 123"
     assert uut._compute_title() is None
 
-    delivery.option_str.return_value = MessageOnlyPolicy.COMBINE_TITLE
-    uut = Envelope(delivery, Notification(mock_context, "testing 123"))
+    delivery.options["message_usage"] = MessageOnlyPolicy.COMBINE_TITLE
+    uut = Envelope(delivery, Notification(ctx, "testing 123"))
     assert uut._compute_message() == "testing 123"
     assert uut._compute_title() is None